feat(Combobox): Debounce `Option` click
