+++ packages/react/src/utilities/useDebounce.ts
@@ -1,0 +1,30 @@
+import { useRef, useEffect } from 'react';
+
+type DebounceFunction<T> = (...args: T[]) => void;
+
+function useDebounce<T>(callback: DebounceFunction<T>, delay = 50) {
+  const timeoutRef = useRef<NodeJS.Timeout | null>(null);
+
+  useEffect(() => {
+    // Cleanup the previous timeout on re-render
+    return () => {
+      if (timeoutRef.current) {
+        clearTimeout(timeoutRef.current);
+      }
+    };
+  }, []);
+
+  const debouncedCallback = (...args: T[]) => {
+    if (timeoutRef.current) {
+      clearTimeout(timeoutRef.current);
+    }
+
+    timeoutRef.current = setTimeout(() => {
+      callback(...args);
+    }, delay);
+  };
+
+  return debouncedCallback;
+}
+
+export default useDebounce;
